mode custom hooks to hooks folder
